move all state logic into Table, make table definitions fully declarative
