+++ packages/network-explorer/src/SessionsTable.tsx
@@ -1,6 +1,4 @@
-import { useState } from "react"
-import useSWR from "swr"
-import { fetchAndIpldParseJson, Result } from "./utils.js"
+import { Result } from "./utils.js"
 import { Action } from "@canvas-js/interfaces"
 import { Table } from "./Table.js"
 import { ColumnDef } from "@tanstack/react-table"
@@ -21,38 +19,11 @@ const defaultColumns: ColumnDef<Result<Action>>[] = [
 	{
 		header: "expiration",
 		accessorKey: "expiration",
-		size: 100,
+		size: 150,
 		enableSorting: false,
 	},
 ]
 
 export const SessionsTable = () => {
-	const [entriesPerPage, setEntriesPerPage] = useState(20)
-	const params: Record<string, string> = {
-		limit: (entriesPerPage + 1).toString(),
-	}
-
-	const { data, mutate: doRefresh } = useSWR(
-		`/api/models/$sessions?${new URLSearchParams(params).toString()}`,
-		fetchAndIpldParseJson<Result<Action>[]>,
-		{
-			refreshInterval: 1000,
-		},
-	)
-
-	const { data: countData } = useSWR(`/api/sessions/count`, fetchAndIpldParseJson<{ count: number }>, {
-		refreshInterval: 1000,
-	})
-
-	return (
-		<Table
-			data={data ? data.content.slice(0, entriesPerPage) : []}
-			rowCount={countData ? countData.content.count : 0}
-			defaultColumns={defaultColumns}
-			entriesPerPage={entriesPerPage}
-			setEntriesPerPage={setEntriesPerPage}
-			responseTime={data?.responseTime}
-			doRefresh={doRefresh}
-		/>
-	)
+	return <Table tableName="$sessions" defaultColumns={defaultColumns} />
 }
